refactor: 최단 경로 책임 이전 및 운임 비용 객체 non-statinc으로 수정
