+++ src/test/java/wooteco/subway/domain/SectionsTest.java
@@ -1,7 +1,6 @@
 package wooteco.subway.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
@@ -37,6 +36,7 @@ class SectionsTest {
         );
     }
 
+    @DisplayName("구간들의 id를 반환한다.")
     @Test
     void getSectionIds() {
         Section section1 = new Section(1L, 1L, 1L, 2L, 10);
@@ -50,6 +50,7 @@ class SectionsTest {
         assertThat(sectionIds).containsExactly(1L, 2L, 3L);
     }
 
+    @DisplayName("2개의 구간을 하나로 합친다.")
     @Test
     void merge() {
         Section section1 = new Section(1L, 1L, 1L, 2L, 10);
@@ -60,63 +61,5 @@ class SectionsTest {
         Section mergedSection = sections.merge();
 
         assertThat(mergedSection.getDistance()).isEqualTo(20);
-    }
-
-    @DisplayName("출발역id와 도착역id를 받아, 최단 경로에 해당하는 지하철역 id들을 반환한다.")
-    @Test
-    void getShortestPathStationIds() {
-        Section section1 = new Section(1L, 1L, 1L, 2L, 2);
-        Section section2 = new Section(2L, 1L, 2L, 3L, 4);
-        Section section3 = new Section(3L, 1L, 3L, 4L, 6);
-        Section section4 = new Section(4L, 2L, 3L, 5L, 8);
-
-        Sections sections = new Sections(List.of(section1, section2, section3, section4));
-
-        List<Long> expected = List.of(1L, 2L, 3L, 5L);
-
-        assertThat(sections.getShortestPathStationIds(1L, 5L)).isEqualTo(expected);
-    }
-
-    @DisplayName("연결되지 않은 구간의 최단 경로를 조회할 경우 예외가 발생한다.")
-    @Test
-    void getShortestPathStationIds_NotConnected() {
-        Section section1 = new Section(1L, 1L, 1L, 2L, 2);
-        Section section2 = new Section(2L, 1L, 2L, 3L, 4);
-        Section section3 = new Section(3L, 1L, 3L, 4L, 6);
-        Section section4 = new Section(4L, 2L, 5L, 6L, 8);
-
-        Sections sections = new Sections(List.of(section1, section2, section3, section4));
-
-        assertThatThrownBy(() -> sections.getShortestPathStationIds(1L, 5L))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("연결되지 않은 구간입니다.");
-    }
-
-    @DisplayName("출발역id와 도착역id를 받아, 최단 거리를 반환한다.")
-    @Test
-    void getShortestPathDistance() {
-        Section section1 = new Section(1L, 1L, 1L, 2L, 2);
-        Section section2 = new Section(2L, 1L, 2L, 3L, 4);
-        Section section3 = new Section(3L, 1L, 3L, 4L, 6);
-        Section section4 = new Section(4L, 2L, 3L, 5L, 8);
-
-        Sections sections = new Sections(List.of(section1, section2, section3, section4));
-
-        assertThat(sections.getShortestPathDistance(1L, 5L)).isEqualTo(14);
-    }
-
-    @DisplayName("연결되지 않은 구간의 최단 경로의 거리를 조회할 경우 예외가 발생한다.")
-    @Test
-    void getShortestPathDistance_NotConnected() {
-        Section section1 = new Section(1L, 1L, 1L, 2L, 2);
-        Section section2 = new Section(2L, 1L, 2L, 3L, 4);
-        Section section3 = new Section(3L, 1L, 3L, 4L, 6);
-        Section section4 = new Section(4L, 2L, 5L, 6L, 8);
-
-        Sections sections = new Sections(List.of(section1, section2, section3, section4));
-
-        assertThatThrownBy(() -> sections.getShortestPathDistance(1L, 5L))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("연결되지 않은 구간입니다.");
     }
 }
